add dispatch for percentiles
