+++ dask/dataframe/backends.py
@@ -1,4 +1,6 @@
 import warnings
+from collections.abc import Iterator
+from functools import wraps
 
 import numpy as np
 import pandas as pd
@@ -27,6 +29,7 @@ from .dispatch import (
     make_meta_dispatch,
     make_meta_obj,
     meta_nonempty,
+    percentile_dispatch,
     tolist_dispatch,
     union_categoricals_dispatch,
 )
@@ -547,3 +550,33 @@ def is_categorical_dtype_pandas(obj):
 @make_meta_obj.register_lazy("cudf")
 def _register_cudf():
     import dask_cudf  # noqa: F401
+
+
+@wraps(np.percentile)
+@percentile_dispatch.register((pd.Series, np.ndarray, pd.Index))
+def percentile_pandas(a, q, interpolation="linear"):
+    n = len(a)
+    if not len(a):
+        return None, n
+    if isinstance(q, Iterator):
+        q = list(q)
+    if a.dtype.name == "category":
+        result = np.percentile(a.cat.codes, q, interpolation=interpolation)
+
+        return pd.Categorical.from_codes(result, a.dtype.categories, a.dtype.ordered), n
+    if isinstance(a.dtype, pd.core.dtypes.dtypes.DatetimeTZDtype) or np.issubdtype(
+        a.dtype, np.datetime64
+    ):
+        if isinstance(a, (pd.Series, pd.Index)):
+            values = a.values
+        else:
+            values = a
+        a2 = values.view("i8")
+        result = np.percentile(a2, q, interpolation=interpolation).astype(values.dtype)
+        if q[0] == 0:
+            # https://github.com/dask/dask/issues/6864
+            result[0] = min(result[0], values.min())
+        return result, n
+    if not np.issubdtype(a.dtype, np.number):
+        interpolation = "nearest"
+    return np.percentile(a, q, interpolation=interpolation), n
