+++ dask/dataframe/dispatch.py
@@ -22,6 +22,12 @@ concat_dispatch = Dispatch("concat")
 tolist_dispatch = Dispatch("tolist")
 is_categorical_dtype_dispatch = Dispatch("is_categorical_dtype")
 union_categoricals_dispatch = Dispatch("union_categoricals")
+percentile_dispatch = Dispatch("percentile_dispatch")
+
+
+def _percentile(a, q, interpolation="linear"):
+    func = percentile_dispatch.dispatch(type(a))
+    return func(a, q, interpolation)
 
 
 def concat(
